+++ src/flask_sqlalchemy/extension.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
 import os
+import types
 import typing as t
+import warnings
 from weakref import WeakKeyDictionary
 
 import sqlalchemy as sa
@@ -14,8 +16,11 @@ from flask import Flask
 from flask import has_app_context
 
 from .model import _QueryProperty
+from .model import BindMixin
 from .model import DefaultMeta
+from .model import DefaultMetaNoName
 from .model import Model
+from .model import NameMixin
 from .pagination import Pagination
 from .pagination import SelectPagination
 from .query import Query
@@ -24,6 +29,33 @@ from .session import Session
 from .table import _Table
 
 _O = t.TypeVar("_O", bound=object)  # Based on sqlalchemy.orm._typing.py
+
+
+# Type accepted for model_class argument
+_FSA_MCT = t.TypeVar(
+    "_FSA_MCT",
+    bound=t.Union[
+        t.Type[Model],
+        sa_orm.DeclarativeMeta,
+        t.Type[sa_orm.DeclarativeBase],
+        t.Type[sa_orm.DeclarativeBaseNoMeta],
+    ],
+)
+
+
+# Type returned by make_declarative_base
+class _FSAModel(Model):
+    metadata: sa.MetaData
+
+
+def _get_2x_declarative_bases(
+    model_class: _FSA_MCT,
+) -> list[t.Type[t.Union[sa_orm.DeclarativeBase, sa_orm.DeclarativeBaseNoMeta]]]:
+    return [
+        b
+        for b in model_class.__bases__
+        if issubclass(b, (sa_orm.DeclarativeBase, sa_orm.DeclarativeBaseNoMeta))
+    ]
 
 
 class SQLAlchemy:
@@ -65,6 +97,18 @@ class SQLAlchemy:
         for a list of arguments.
     :param add_models_to_shell: Add the ``db`` instance and all model classes to
         ``flask shell``.
+
+    .. versionchanged:: 3.1.0
+        The ``metadata`` parameter can still be used with SQLAlchemy 1.x classes,
+        but is ignored when using SQLAlchemy 2.x style of declarative classes.
+        Instead, specify metadata on your Base class.
+
+    .. versionchanged:: 3.1.0
+        Added the ``disable_autonaming`` parameter.
+
+    .. versionchanged:: 3.1.0
+        Changed ``model_class`` parameter to accepta SQLAlchemy 2.x
+        declarative base subclass.
 
     .. versionchanged:: 3.0
         An active Flask application context is always required to access ``session`` and
@@ -124,9 +168,10 @@ class SQLAlchemy:
         metadata: sa.MetaData | None = None,
         session_options: dict[str, t.Any] | None = None,
         query_class: type[Query] = Query,
-        model_class: type[Model] | sa_orm.DeclarativeMeta = Model,
+        model_class: _FSA_MCT = Model,  # type: ignore[assignment]
         engine_options: dict[str, t.Any] | None = None,
         add_models_to_shell: bool = True,
+        disable_autonaming: bool = False,
     ):
         if session_options is None:
             session_options = {}
@@ -168,8 +213,17 @@ class SQLAlchemy:
         """
 
         if metadata is not None:
-            metadata.info["bind_key"] = None
-            self.metadatas[None] = metadata
+            if len(_get_2x_declarative_bases(model_class)) > 0:
+                warnings.warn(
+                    "When using SQLAlchemy 2.x style of declarative classes,"
+                    " the `metadata` should be an attribute of the base class."
+                    "The metadata passed into SQLAlchemy() is ignored.",
+                    DeprecationWarning,
+                    stacklevel=2,
+                )
+            else:
+                metadata.info["bind_key"] = None
+                self.metadatas[None] = metadata
 
         self.Table = self._make_table_class()
         """A :class:`sqlalchemy.schema.Table` class that chooses a metadata
@@ -187,7 +241,9 @@ class SQLAlchemy:
             This is a subclass of SQLAlchemy's ``Table`` rather than a function.
         """
 
-        self.Model = self._make_declarative_base(model_class)
+        self.Model = self._make_declarative_base(
+            model_class, disable_autonaming=disable_autonaming
+        )
         """A SQLAlchemy declarative model class. Subclass this to define database
         models.
 
@@ -199,9 +255,15 @@ class SQLAlchemy:
         database engine. Otherwise, it will use the default :attr:`metadata` and
         :attr:`engine`. This is ignored if the model sets ``metadata`` or ``__table__``.
 
-        Customize this by subclassing :class:`.Model` and passing the ``model_class``
-        parameter to the extension. A fully created declarative model class can be
+        For code using the SQLAlchemy 1.x API, customize this model by subclassing
+        :class:`.Model` and passing the ``model_class`` parameter to the extension.
+        A fully created declarative model class can be
         passed as well, to use a custom metaclass.
+
+        For code using the SQLAlchemy 2.x API, customize this model by subclassing
+        :class:`sqlalchemy.orm.DeclarativeBase` or
+        :class:`sqlalchemy.orm.DeclarativeBaseNoMeta`
+        and passing the ``model_class`` parameter to the extension.
         """
 
         if engine_options is None:
@@ -439,8 +501,10 @@ class SQLAlchemy:
         return Table
 
     def _make_declarative_base(
-        self, model: type[Model] | sa_orm.DeclarativeMeta
-    ) -> type[t.Any]:
+        self,
+        model_class: _FSA_MCT,
+        disable_autonaming: bool = False,
+    ) -> t.Type[_FSAModel]:
         """Create a SQLAlchemy declarative model class. The result is available as
         :attr:`Model`.
 
@@ -452,7 +516,14 @@ class SQLAlchemy:
 
         :meta private:
 
-        :param model: A model base class, or an already created declarative model class.
+        :param model_class: A model base class, or an already created declarative model
+        class.
+
+        :param disable_autonaming: Turns off automatic tablename generation in models.
+
+        .. versionchanged:: 3.1.0
+            Added support for passing SQLAlchemy 2.x base class as model class.
+            Added optional ``disable_autonaming`` parameter.
 
         .. versionchanged:: 3.0
             Renamed with a leading underscore, this method is internal.
@@ -460,22 +531,45 @@ class SQLAlchemy:
         .. versionchanged:: 2.3
             ``model`` can be an already created declarative model class.
         """
-        if not isinstance(model, sa_orm.DeclarativeMeta):
-            metadata = self._make_metadata(None)
-            model = sa_orm.declarative_base(
-                metadata=metadata, cls=model, name="Model", metaclass=DefaultMeta
+        model: t.Type[_FSAModel]
+        declarative_bases = _get_2x_declarative_bases(model_class)
+        if len(declarative_bases) > 1:
+            # raise error if more than one declarative base is found
+            raise ValueError(
+                "Only one declarative base can be passed to SQLAlchemy."
+                " Got: {}".format(model_class.__bases__)
             )
+        elif len(declarative_bases) == 1:
+            body = dict(model_class.__dict__)
+            body["__fsa__"] = self
+            mixin_classes = [BindMixin, NameMixin, Model]
+            if disable_autonaming:
+                mixin_classes.remove(NameMixin)
+            model = types.new_class(
+                "FlaskSQLAlchemyBase",
+                (*mixin_classes, *model_class.__bases__),
+                {"metaclass": type(declarative_bases[0])},
+                lambda ns: ns.update(body),
+            )
+        elif not isinstance(model_class, sa_orm.DeclarativeMeta):
+            metadata = self._make_metadata(None)
+            metaclass = DefaultMetaNoName if disable_autonaming else DefaultMeta
+            model = sa_orm.declarative_base(
+                metadata=metadata, cls=model_class, name="Model", metaclass=metaclass
+            )
+        else:
+            model = model_class  # type: ignore[assignment]
 
         if None not in self.metadatas:
             # Use the model's metadata as the default metadata.
-            model.metadata.info["bind_key"] = None  # type: ignore[union-attr]
-            self.metadatas[None] = model.metadata  # type: ignore[union-attr]
+            model.metadata.info["bind_key"] = None
+            self.metadatas[None] = model.metadata
         else:
             # Use the passed in default metadata as the model's metadata.
-            model.metadata = self.metadatas[None]  # type: ignore[union-attr]
+            model.metadata = self.metadatas[None]
 
         model.query_class = self.Query
-        model.query = _QueryProperty()
+        model.query = _QueryProperty()  # type: ignore[assignment]
         model.__fsa__ = self
         return model